Refactored OperatorExecutors
